Add the ability to write a single row
Bringing the ability to write a new row up a level allows the code to be
a little cleaner, and exposes one more piece to the outside. This allows
users to be able to change pieces of a bitmap instead of the entire
board.

Amos King @adkron <amos@binarynoggin.com>