Fixed unexpected modification of input bitmap
When writeBitmap is called "row.unshift(row.pop());" modifies the array
object that is passed to writeBitmap. If the bitmap is only used once it
is OK, however if the bitmap is to be reused it is unexpectedly shifted
each time writeBitmap is called. This commit prevents this from
happening.